common: Move PackageNEVRA to common
Related: RHEL-60139
